Enhance course DTOs and queries for user-specific data
- Add user-specific fields (IsEnrolled, Progress, etc.) to CourseResponse DTO
- Change ProgressPercentage type to int in CourseDetailResponse
- Introduce GetAllCoursesRequest to separate API and MediatR models
- Update query handler to compute user-specific course data
- Adjust controller to use new request model and pass UserID
- Fix ProgressPercentage calculation in repository to match new type
- Improve pagination variable naming and logic

+++ Application/DTOs/Courses/CourseDetailResponse.cs
@@ -15,7 +15,7 @@ namespace Application.DTOs.Courses
         public int NumberOfSheets { get; set; }
         public int NumberOfWatchedVideos { get; set; }
         public int NumberOfSections { get; set; }
-        public decimal ProgressPercentage { get; set; }
+        public int ProgressPercentage { get; set; }
         public DateTimeOffset CreatedAt { get; set; }
         public DateTimeOffset UpdatedAt { get; set; }
         public List<InstructorInfoDto> Instructors { get; set; } = [];
